+++ .idea/misc.xml
@@ -5,8 +5,9 @@
       <map>
         <entry key="..\:/Users/rulia/AndroidStudioProjects/Calculator/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.14479166666666668" />
         <entry key="..\:/Users/rulia/AndroidStudioProjects/Calculator/app/src/main/res/drawable/ic_launcher_background.xml" value="0.14479166666666668" />
-        <entry key="..\:/Users/rulia/AndroidStudioProjects/Calculator/app/src/main/res/layout-land/activity_main.xml" value="0.4" />
-        <entry key="..\:/Users/rulia/AndroidStudioProjects/Calculator/app/src/main/res/layout/activity_main.xml" value="0.4" />
+        <entry key="..\:/Users/rulia/AndroidStudioProjects/Calculator/app/src/main/res/layout-land/activity_main.xml" value="0.264" />
+        <entry key="..\:/Users/rulia/AndroidStudioProjects/Calculator/app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="..\:/Users/rulia/AndroidStudioProjects/Calculator/app/src/main/res/layout/activity_settings.xml" value="0.2" />
       </map>
     </option>
   </component>
