+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".SettingsActivity">
+
+    </activity>
     </application>
 
 </manifest>
