+++ app/src/main/java/ru/evgenysidorov/calculator/Constants.java
@@ -1,0 +1,5 @@
+package ru.evgenysidorov.calculator;
+
+public interface Constants {
+    String YOUR_STYLE = "YOUR_STYLE";
+}
